requirements.txt
django
django-registration-redux
django-crispy-forms
pylint-django
django-progressive-web-app
mysqlclient
Pillow
django-import-export